feat(agent): 添加租户配额配置及配额校验支持
- 在 AgentProperties 中新增租户配额配置子类 Quota，支持每日 Token 限额和月度预算等参数设置
- ChatService 集成租户配额管理服务，调用前进行配额预检，调用后记录实际用量
- 解析当前调用的租户 ID，默认 fallback 为 "default"
- 新增 TenantQuota 对象封装配置信息，便于业务调用使用
- 优化 Cronjob 模块部分命名及 Redis 操作，确保代码规范一致
- 改用 LambdaQueryWrapper 替换部分 Repository 中老旧 QueryWrapper，提升代码类型安全性
- 完善内置表达式函数分类及注册，明确函数归属类别
- ChunkUpload 服务中切换配额调用为存储配额仓库，修正保存文件节点类型
- 添加配置接口入参字段非空校验注解，提升接口健壮性
- 修正告警相关对象中字段命名，避免歧义与冲突
- DagInstanceExecutor 新增判断与剥离 LOOP 迭代后缀工具方法
- 调整重复定义内部类 RefreshRequest，移除冗余代码
- 优化 Map.of 返回类型，避免泛型推断警告

+++ ydsz-cronjob/ydsz-cronjob-server/src/main/java/com/njydsz/cronjob/server/core/GracefulShutdownCoordinator.java
@@ -160,7 +160,7 @@ public class GracefulShutdownCoordinator implements SmartLifecycle {
      * <p>轮询等待运行中任务数降为 0，超时后强制继续。
      */
     private void drainRunningTasks() {
-        int drainTimeout = cronjobProperties.getExecutor().getDrainTimeoutSeconds();
+        long drainTimeout = cronjobProperties.getExecutor().getDrainTimeoutSeconds();
         log.info("[GracefulShutdown] 阶段2: 等待运行中任务完成, 超时={}s, 当前运行中={}",
                 drainTimeout, runningTaskCount.get());
